+++ podaac/hitide_backfill_tool/cli.py
@@ -10,6 +10,7 @@ import uuid
 import copy
 import json
 from datetime import datetime, timezone
+import traceback
 import requests
 
 from podaac.hitide_backfill_tool.cmr.search import GranuleSearch
@@ -79,11 +80,14 @@ def granule_search_from_args(args, logger):
 def message_writer_from_args(args, logger):
     """Return configured message writer from parsed cli args and logger."""
 
-    message_config = get_message_config(args.cumulus, args.default_message_config)
-    collection_config = get_collection_config(
-        args.cumulus_configurations, args.collection, args.cumulus, logger)
-    message_writer = CnmMessageWriter(message_config, collection_config,
-                                      args.start_date, args.end_date, args.provider, args.cli_execution_id, args.user)
+    message_writer = None
+
+    if args.cumulus_configurations and args.default_message_config:
+        message_config = get_message_config(args.cumulus, args.default_message_config)
+        collection_config = get_collection_config(
+            args.cumulus_configurations, args.collection, args.cumulus, logger)
+        message_writer = CnmMessageWriter(message_config, collection_config,
+                                          args.start_date, args.end_date, args.provider, args.cli_execution_id, args.user)
     return message_writer
 
 
@@ -123,11 +127,10 @@ class Backfiller:
     # Disable broad-except since many types of error indicate the absense
     #   of data when attempting access (e.g. TypeError, IndexError, KeyError, ...)
     # pylint: disable=broad-except
-
-    # pylint: disable=too-many-instance-attributes,too-many-arguments
+    # pylint: disable=too-many-instance-attributes,too-many-arguments,too-many-positional-arguments
 
     def __init__(self, search, message_writer, message_senders, granule_options, logger,
-                 message_limit, cli_execution_id, s3, collection):
+                 message_limit, cli_execution_id, s3, collection, granule_list_file):
         # pylint: disable=C0103
 
         # dependencies
@@ -140,6 +143,8 @@ class Backfiller:
         self.cli_execution_id = cli_execution_id
         self.s3 = s3
         self.collection = collection
+        self.granule_list_file = granule_list_file
+        self.granule_list = None
 
         # statistics
         self.granules_analyzed = 0
@@ -178,27 +183,52 @@ class Backfiller:
             destination_message.append('nowhere')
         self.destination_message = f"Messages being sent to {', '.join(destination_message)}"
 
+        if self.granule_list_file:
+            self.read_granule_list_file()
+
         # for thread-safe operations
         self.lock = Lock()
+
+    def read_granule_list_file(self):
+        """Read the granule_list_file and store contents in array granule_list
+           The items in the list are one granule per line:
+           Can be either a GranuleUR or a granlue concept ID, and all items must the same type.
+           Example:
+              20240702060501-JPL-L2P_GHRSST-SSTskin-MODIS_A-D-v02.0-fv01.0
+              20240702083501-JPL-L2P_GHRSST-SSTskin-MODIS_A-N-v02.0-fv01.0
+              20240702174000-JPL-L2P_GHRSST-SSTskin-MODIS_A-D-v02.0-fv01.0
+           OR
+               G3141860732-POCLOUD
+               G3142056458-POCLOUD
+               G3142846484-POCLOUD
+        """
+
+        with open(self.granule_list_file, 'r', encoding='utf-8') as file:
+            self.granule_list = [line.strip() for line in file]
 
     def process_granules(self):
         """Loop through granules (in parallel) from granule-search and call the process_one_granule() method."""
 
-        while self.search.get_next_page():
-            print("Processing granules...", end='', flush=True)
+        if self.granule_list:
+            print('Processing granules from granule list file...', end='', flush=True)
+            granules = self.search.get_granules_in_list(self.granule_list)
+
             with ThreadPoolExecutor() as executor:
-                executor.map(self.process_one_granule, self.search.granules())
+                executor.map(self.process_one_granule, granules)
             print("done.")
-            if self.message_limit_reached():
-                self.logger.info("\n**** Message limit reached ****")
-                return
-            self.log_stats()
+        else:
+            while self.search.get_next_page():
+                print("Processing granules...", end='', flush=True)
+                with ThreadPoolExecutor() as executor:
+                    executor.map(self.process_one_granule, self.search.granules())
+                print("done.")
+                if self.message_limit_reached():
+                    self.logger.info("\n**** Message limit reached ****")
+                    return
+                self.log_stats()
 
     def print_monthly_results_table(self):
         """Function to print out monthly stats"""
-
-        if not self.message_senders:
-            print("** NOTE: When in preview mode, the messages sent count may not be accurate since it's only simulating sending messages. ** \n")
 
         print("Monthly Counts Summary:\n")
         header = f"{'Date':<10} {'Granules':<10} {'Need Image':<12} {'Need Footprint':<16} {'Both FP & BBox':<16} {'Need DMRPP':<12}"
@@ -274,6 +304,7 @@ class Backfiller:
                 self.granules_analyzed += 1
         except Exception as exc:
             self.logger.error(f"Error: {str(exc)}\n")
+            traceback.print_exc()
 
     def update_image(self, granule):
         """Create and send messages for one granule's image update."""
@@ -288,11 +319,12 @@ class Backfiller:
             if not self.message_limit_reached():
                 with self.lock:
                     self.image_messages_sent += 1
-                message = self.message_writer.write(granule, needs_footprint=False,
-                                                    needs_image=True, needs_dmrpp=False,
-                                                    skip_cmr_opendap_update=True)
-                for sender in self.message_senders:
-                    sender.send(message)
+                if self.message_writer:
+                    message = self.message_writer.write(granule, needs_footprint=False,
+                                                        needs_image=True, needs_dmrpp=False,
+                                                        skip_cmr_opendap_update=True)
+                    for sender in self.message_senders:
+                        sender.send(message)
         else:
             with self.lock:
                 self.images_that_couldnt_be_processed += 1
@@ -313,11 +345,12 @@ class Backfiller:
             if not self.message_limit_reached():
                 with self.lock:
                     self.footprint_messages_sent += 1
-                message = self.message_writer.write(granule, needs_footprint=True,
-                                                    needs_image=False, needs_dmrpp=False,
-                                                    skip_cmr_opendap_update=True)
-                for sender in self.message_senders:
-                    sender.send(message)
+                if self.message_writer:
+                    message = self.message_writer.write(granule, needs_footprint=True,
+                                                        needs_image=False, needs_dmrpp=False,
+                                                        skip_cmr_opendap_update=True)
+                    for sender in self.message_senders:
+                        sender.send(message)
         else:
             with self.lock:
                 self.footprints_that_couldnt_be_processed += 1
@@ -330,22 +363,25 @@ class Backfiller:
 
         s3_bucket_info = granule.s3_bucket_info()
         if s3_bucket_info:
-            dmrpp_state = granule.get_dmrpp_state(f's3://{s3_bucket_info["bucket"]}'
-                                                  f'/{s3_bucket_info["key"]}.dmrpp')
-            if dmrpp_state == DmrppState.OLDER_VERSION:
+            if not granule.has_opendap_url():
                 self.update_dmrpp(granule)
-                with self.lock:
-                    self.dmrpp_older_version += 1
-            elif dmrpp_state == DmrppState.MISSING_VERSION:
-                self.update_dmrpp(granule)
-                with self.lock:
-                    self.dmrpp_missing_version += 1
-            elif dmrpp_state == DmrppState.MATCHED_VERSION:
-                with self.lock:
-                    self.dmrpp_unprocessed += 1
-            elif dmrpp_state == DmrppState.NEWER_VERSION:
-                with self.lock:
-                    self.dmrpp_newer_version += 1
+            else:
+                dmrpp_state = granule.get_dmrpp_state(f's3://{s3_bucket_info["bucket"]}'
+                                                      f'/{s3_bucket_info["key"]}.dmrpp')
+                if dmrpp_state == DmrppState.OLDER_VERSION:
+                    self.update_dmrpp(granule)
+                    with self.lock:
+                        self.dmrpp_older_version += 1
+                elif dmrpp_state == DmrppState.MISSING_VERSION:
+                    self.update_dmrpp(granule)
+                    with self.lock:
+                        self.dmrpp_missing_version += 1
+                elif dmrpp_state == DmrppState.MATCHED_VERSION:
+                    with self.lock:
+                        self.dmrpp_unprocessed += 1
+                elif dmrpp_state == DmrppState.NEWER_VERSION:
+                    with self.lock:
+                        self.dmrpp_newer_version += 1
         else:
             with self.lock:
                 self.dmrpp_that_couldnt_be_processed += 1
@@ -370,11 +406,12 @@ class Backfiller:
             if not self.message_limit_reached():
                 with self.lock:
                     self.dmrpp_messages_sent += 1
-                message = self.message_writer.write(granule, needs_footprint=False,
-                                                    needs_image=False, needs_dmrpp=True,
-                                                    skip_cmr_opendap_update=skip_cmr_opendap_update)
-                for sender in self.message_senders:
-                    sender.send(message)
+                if self.message_writer:
+                    message = self.message_writer.write(granule, needs_footprint=False,
+                                                        needs_image=False, needs_dmrpp=True,
+                                                        skip_cmr_opendap_update=skip_cmr_opendap_update)
+                    for sender in self.message_senders:
+                        sender.send(message)
         else:
             with self.lock:
                 self.dmrpp_that_couldnt_be_processed += 1
@@ -420,6 +457,9 @@ class Backfiller:
             f"-- {self.destination_message} --\n"
             "==============================================================\n"
         )
+        if not self.message_senders:
+            print("** NOTE: When in preview mode, the messages sent count may not be accurate since it's only simulating sending messages. ** \n")
+
         if len(self.concept_ids_needing_image) > 0:
             self.logger.info(f"Granule IDs needing images (showing first 100):\n"
                              f" {self.concept_ids_needing_image[:100]}\n"
@@ -452,6 +492,74 @@ class Backfiller:
             self.forge_tig_configuration = None
 
 
+def validate_arg(name, value, allowed_values):
+    """Raises an exception if the value is invalid."""
+    allowed_values_str = ", ".join(f"'{v}'" for v in allowed_values[:-1]) + f", or '{allowed_values[-1]}'"
+    if value is None:
+        raise Exception(f"Please specify --{name}.  Must be either {allowed_values_str}")
+    if value not in allowed_values:
+        raise Exception(f"Invalid value for --{name}.  Must be either {allowed_values_str}")
+
+
+def verify_inputs(args, granule_options, message_writer, backfiller):
+    """Verify inputs from parsed cli args, and raise an exception if any are invalid."""
+
+    # pylint: disable=too-many-branches
+
+    allowed_values = ['off', 'on', 'force']
+
+    validate_arg('image', args.image, allowed_values)
+    validate_arg('footprint', args.footprint, allowed_values)
+    validate_arg('dmrpp', args.dmrpp, allowed_values)
+
+    if args.image == "off" and args.footprint == "off" and args.dmrpp == "off":
+        raise Exception("At least one of --image, --footprint, or --dmrpp must be 'on' or 'force'")
+
+    if args.default_message_config is None and not args.preview:
+        raise Exception("Please specify path to default message config file")
+
+    if args.cumulus_configurations is None and not args.preview:
+        raise Exception("Please specify path to cumulus configurations directory")
+
+    # Check forge configurations before running backfill
+    backfiller.get_forge_tig_configuration()
+
+    if granule_options['footprint_processing'] != "off":
+        if backfiller.forge_tig_configuration is None:
+            raise Exception("Cannot find forge-tig configuration for this collection")
+        footprint_settings = backfiller.forge_tig_configuration.get('footprint')
+        if not footprint_settings:
+            raise Exception("There is no footprint setting for this collection, please disable footprint for backfilling")
+
+    if granule_options['image_processing'] != "off":
+        if backfiller.forge_tig_configuration is None:
+            raise Exception("Cannot find forge-tig configuration for this collection")
+        image_settings = backfiller.forge_tig_configuration.get('imgVariables')
+        if not image_settings:
+            raise Exception("There is no image setting for this collection, please disable image for backfilling")
+
+    if granule_options['dmrpp_processing'] != "off":
+        if message_writer is None:
+            error_msg = "Dmrpp processing requires "
+            missing = []
+            if args.cumulus_configurations is None:
+                missing.append("cumulus configurations directory")
+            if args.default_message_config is None:
+                missing.append("default message config file")
+
+            error_msg += " and ".join(missing) + " to be specified" if missing else ""
+            raise Exception(error_msg)
+
+        files = message_writer.collection_config.get('files', [])
+        has_dmrpp_regex = False
+        for file in files:
+            if file.get('regex', "").endswith(".dmrpp$"):
+                has_dmrpp_regex = True
+                break
+        if has_dmrpp_regex is False:
+            raise Exception(f"There is no DMRPP regex in cumulus collection configuration for {message_writer.collection_name}")
+
+
 def main(args=None):
     """Main script for backfilling from the cli"""
 
@@ -475,7 +583,7 @@ def main(args=None):
         message_writer = message_writer_from_args(args, logger)
         message_senders = message_senders_from_args(args, logger)
         granule_options = granule_options_from_args(args)
-        s3 = S3Reader(logger, args.aws_profile)    # pylint: disable=C0103
+        s3 = S3Reader(logger, args.aws_profile)
         collection = args.collection
     except Exception as exc:
         logger.error(f"Error: {str(exc)}\n")
@@ -483,27 +591,13 @@ def main(args=None):
 
     # setup backfiller
     backfiller = Backfiller(search, message_writer, message_senders,
-                            granule_options, logger, args.message_limit, args.cli_execution_id, s3, collection)
+                            granule_options, logger, args.message_limit, args.cli_execution_id, s3, collection, args.granule_list_file)
 
-    # Check forge configurations before running backfill
-    backfiller.get_forge_tig_configuration()
-
-    if granule_options['footprint_processing'] != "off":
-        if backfiller.forge_tig_configuration is None:
-            raise Exception("There is no footprint settings for this collection, please disable footprint for backfilling")
-        footprint_settings = backfiller.forge_tig_configuration.get('footprint')
-        if not footprint_settings:
-            raise Exception("There is no footprint settings for this collection, please disable footprint for backfilling")
-
-    if granule_options['dmrpp_processing'] != "off":
-        files = message_writer.collection_config.get('files', [])
-        has_dmrpp_regex = False
-        for file in files:
-            if file.get('regex', "").endswith(".dmrpp$"):
-                has_dmrpp_regex = True
-                break
-        if has_dmrpp_regex is False:
-            raise Exception(f"There is no DMRPP regex in cumulus collection configuration for {message_writer.collection_name}")
+    try:
+        verify_inputs(args, granule_options, message_writer, backfiller)
+    except Exception as exc:
+        logger.error(f"Error: {str(exc)}\n")                              # pylint: disable=W1203
+        return
 
     # run backfiller
     try: