+++ terraform-deploy/forge_py.tf
@@ -10,6 +10,20 @@ module "forge_py_module" {
     lambda_role = aws_iam_role.iam_execution.arn
     security_group_ids = [var.aws_security_group_ids]
     subnet_ids = data.aws_subnets.private.ids
-    memory_size = 1024
+    memory_size = 2048
     timeout = 900
+
+    # ECS Variables
+
+    # Fargate Variables
+    ecs_cluster_arn = aws_ecs_cluster.main.arn
+    forge_py_fargate = true
+    fargate_iam_role = aws_iam_role.fargate_execution.arn
+    ecs_cluster_name = aws_ecs_cluster.main.name
+    fargate_max_capacity = 100
+    fargate_cpu = 2048
+    fargate_memory = 16384
+    threshold_scale_down = 25
+    period_scale_down = 240
+
 }