+++ terraform-deploy/post_to_cmr.tf
@@ -1,9 +1,10 @@
 resource "aws_lambda_function" "post_to_cmr_task" {
   function_name    = "${local.resources_name}-PostToCmr"
   filename         = "post_to_cmr.zip"
+  source_code_hash = filebase64sha256("post_to_cmr.zip")
   handler          = "index.handler"
   role             = aws_iam_role.iam_execution.arn
-  runtime          = "nodejs16.x"
+  runtime          = var.cumulus_node_version
   timeout          = 300
   memory_size      = 512
 
@@ -23,9 +24,4 @@ resource "aws_lambda_function" "post_to_cmr_task" {
       CUMULUS_MESSAGE_ADAPTER_DIR = "/opt/"
     }
   }
-
-  tags = {
-    Version = "18.2.0"
-  }
-
 }