+++ terraform-deploy/tig_step.tf
@@ -6,6 +6,7 @@ resource "aws_sfn_state_machine" "tig" {
 {
   "Comment": "Image Processing",
   "StartAt": "Normalizer",
+  "TimeoutSeconds": 10800,
   "States": {
     "Normalizer": {
       "Parameters": {
@@ -50,25 +51,12 @@ resource "aws_sfn_state_machine" "tig" {
         {
           "And": [
             {
-              "Variable": "$.meta.collection.meta.workflowChoice.useEcs",
+              "Variable": "$.meta.collection.meta.workflowChoice.ecs_lambda",
               "IsPresent": true
             },
             {
-              "Variable": "$.meta.collection.meta.workflowChoice.useEcs",
-              "BooleanEquals": true
-            }
-          ],
-          "Next": "ECSImageProcess"
-        },
-        {
-          "And": [
-            {
-              "Variable": "$.meta.collection.meta.workflowChoice.image_ecs",
-              "IsPresent": true
-            },
-            {
-              "Variable": "$.meta.collection.meta.workflowChoice.image_ecs",
-              "BooleanEquals": true
+              "Variable": "$.meta.collection.meta.workflowChoice.ecs_lambda",
+              "StringEquals": "ecs"
             }
           ],
           "Next": "ECSImageProcess"